requirements.txt
fastapi
uvicorn
httpx
langchain
langgraph
pydantic
faiss-cpu